알림 ( 스케쥴러 설정 ) 완료
모니터링 on 되어있는 sensor name
불러와서 마지막으로 업데이트 된 데이터를
기준값과 비교해서 높은경우 알림.

추가되어야할것
> 날짜 비교해서 최근 시간보다 10분 초과된 경우, 상태 값 OFF 로 알림보내야할듯

+++ src/main/java/com/example/tms/controller/JsonController.java
@@ -128,32 +128,32 @@ public class JsonController {
     }
     //법적기준
     @RequestMapping(value = "/getLegal")
-    public Double getLegal(@RequestParam("name") String name){
+    public Float getLegal(@RequestParam("name") String name){
         try{
             return reference_value_settingRepository.findByName(name).getLegal_standard();
 
         }catch (NullPointerException e){
-            return 0.0;
+            return 0.0f;
         }
     }
     //사내기준
     @RequestMapping(value = "/getCompany")
-    public Double getCompany(@RequestParam("name") String name){
+    public Float getCompany(@RequestParam("name") String name){
         try{
             return reference_value_settingRepository.findByName(name).getCompany_standard();
 
         }catch (NullPointerException e){
-            return 0.0;
+            return 0.0f;
         }
     }
     //관리기준
     @RequestMapping(value = "/getManagement")
-    public Double getManagement(@RequestParam("name") String name){
+    public Float getManagement(@RequestParam("name") String name){
         try{
             return reference_value_settingRepository.findByName(name).getManagement_standard();
 
         }catch (NullPointerException e){
-            return 0.0;
+            return 0.0f;
         }
     }
